+++ datadog_checks_dev/datadog_checks/dev/tooling/templates/integration/jmx/{check_name}/datadog_checks/{check_name}/config_models/shared.py
@@ -4,7 +4,8 @@
 
 from __future__ import annotations
 
-from typing import Any, Mapping, Optional, Sequence
+from types import MappingProxyType
+from typing import Any, Optional
 
 from pydantic import BaseModel, ConfigDict, field_validator, model_validator
 
@@ -17,10 +18,11 @@ from . import defaults, validators
 class SharedConfig(BaseModel):
     model_config = ConfigDict(
         validate_default=True,
+        arbitrary_types_allowed=True,
         frozen=True,
     )
     collect_default_metrics: bool
-    conf: Optional[Sequence[Mapping[str, Any]]] = None
+    conf: Optional[tuple[MappingProxyType[str, Any], ...]] = None
     is_jmx: bool
     new_gc_metrics: Optional[bool] = None
     service: Optional[str] = None
@@ -31,25 +33,14 @@ class SharedConfig(BaseModel):
         return validation.core.initialize_config(getattr(validators, 'initialize_shared', identity)(values))
 
     @field_validator('*', mode='before')
-    def _ensure_defaults(cls, value, info):
+    def _validate(cls, value, info):
         field = cls.model_fields[info.field_name]
         field_name = field.alias or info.field_name
         if field_name in info.context['configured_fields']:
-            return value
+            value = getattr(validators, f'shared_{{info.field_name}}', identity)(value, field=field)
+        else:
+            value = getattr(defaults, f'shared_{{info.field_name}}', lambda: value)()
 
-        return getattr(defaults, f'shared_{{info.field_name}}', lambda: value)()
-
-    @field_validator('*')
-    def _run_validations(cls, value, info):
-        field = cls.model_fields[info.field_name]
-        field_name = field.alias or info.field_name
-        if field_name not in info.context['configured_fields']:
-            return value
-
-        return getattr(validators, f'shared_{{info.field_name}}', identity)(value, field=field)
-
-    @field_validator('*', mode='after')
-    def _make_immutable(cls, value):
         return validation.utils.make_immutable(value)
 
     @model_validator(mode='after')