+++ postgres/datadog_checks/postgres/metrics_cache.py
@@ -93,9 +93,10 @@ class PostgresMetricsCache:
             'relation': False,
         }
 
-        res["query"] += " WHERE " + " AND ".join(
-            "psd.datname not ilike '{}'".format(db) for db in self.config.ignore_databases
-        )
+        if len(self.config.ignore_databases) > 0:
+            res["query"] += " WHERE " + " AND ".join(
+                "psd.datname not ilike '{}'".format(db) for db in self.config.ignore_databases
+            )
 
         if self.config.dbstrict:
             res["query"] += " AND psd.datname in('{}')".format(self.config.dbname)