+++ snmp/tests/test_e2e_core_profiles/test_profile_nvidia_mellanox_switchx.py
@@ -10,7 +10,7 @@ from .. import common
 from ..test_e2e_core_metadata import assert_device_metadata
 from .utils import (
     assert_common_metrics,
-    assert_extend_entity_sensor,
+    assert_extend_generic_entity_sensor,
     assert_extend_generic_host_resources,
     assert_extend_generic_if,
     create_e2e_core_test_config,
@@ -35,7 +35,7 @@ def test_e2e_profile_nvidia_mellanox_switchx(dd_agent_check):
     # --- TEST METRICS ---
     assert_common_metrics(aggregator, common_tags)
     assert_extend_generic_if(aggregator, common_tags)
-    assert_extend_entity_sensor(aggregator, common_tags)
+    assert_extend_generic_entity_sensor(aggregator, common_tags)
     assert_extend_generic_host_resources(aggregator, common_tags)
     aggregator.assert_all_metrics_covered()