+++ snowflake/datadog_checks/snowflake/config_models/shared.py
@@ -9,7 +9,8 @@
 
 from __future__ import annotations
 
-from typing import Any, Mapping, Optional, Sequence
+from types import MappingProxyType
+from typing import Any, Optional
 
 from pydantic import BaseModel, ConfigDict, field_validator, model_validator
 
@@ -22,9 +23,10 @@ from . import validators
 class SharedConfig(BaseModel):
     model_config = ConfigDict(
         validate_default=True,
+        arbitrary_types_allowed=True,
         frozen=True,
     )
-    global_custom_queries: Optional[Sequence[Mapping[str, Any]]] = None
+    global_custom_queries: Optional[tuple[MappingProxyType[str, Any], ...]] = None
     proxy_host: Optional[str] = None
     proxy_password: Optional[str] = None
     proxy_port: Optional[int] = None
@@ -35,17 +37,13 @@ class SharedConfig(BaseModel):
     def _initial_validation(cls, values):
         return validation.core.initialize_config(getattr(validators, 'initialize_shared', identity)(values))
 
-    @field_validator('*')
-    def _run_validations(cls, value, info):
+    @field_validator('*', mode='before')
+    def _validate(cls, value, info):
         field = cls.model_fields[info.field_name]
         field_name = field.alias or info.field_name
-        if field_name not in info.context['configured_fields']:
-            return value
+        if field_name in info.context['configured_fields']:
+            value = getattr(validators, f'shared_{info.field_name}', identity)(value, field=field)
 
-        return getattr(validators, f'shared_{info.field_name}', identity)(value, field=field)
-
-    @field_validator('*', mode='after')
-    def _make_immutable(cls, value):
         return validation.utils.make_immutable(value)
 
     @model_validator(mode='after')