+++ vsphere/datadog_checks/vsphere/constants.py
@@ -66,6 +66,13 @@ PROPERTY_COUNT_METRICS = [
     "guest.toolsVersionStatus2",
     "guest.toolsVersion",
     "guest.guestFullName",
+    "hardware.cpuPowerManagementInfo.currentPolicy",
+    "summary.runtime.connectionState",
+    "summary.runtime.powerState",
+    "summary.runtime.inMaintenanceMode",
+    "configuration.drsConfig.enabled",
+    "configuration.drsConfig.defaultVmBehavior",
+    "configuration.dasConfig.enabled",
 ]
 VM_OBJECT_PROPERTIES = ["guest.disk", "guest.net", "guest.ipStack"]
 
@@ -86,4 +93,46 @@ VM_SIMPLE_PROPERTIES = [
     "guest.guestFullName",
 ]
 
+HOST_SIMPLE_PROPERTIES = [
+    "hardware.cpuPowerManagementInfo.currentPolicy",
+    "summary.runtime.connectionState",
+    "summary.runtime.powerState",
+    "summary.runtime.inMaintenanceMode",
+]
+
+CLUSTER_SIMPLE_PROPERTIES = [
+    "configuration.drsConfig.enabled",
+    "configuration.drsConfig.defaultVmBehavior",
+    "configuration.drsConfig.vmotionRate",
+    "configuration.dasConfig.enabled",
+]
+
+
+DATASTORE_SIMPLE_PROPERTIES = [
+    "summary.capacity",
+    "summary.freeSpace",
+]
+
+DATASTORE_PROPERTIES = DATASTORE_SIMPLE_PROPERTIES
+
+CLUSTER_PROPERTIES = CLUSTER_SIMPLE_PROPERTIES
+
+HOST_PROPERTIES = HOST_SIMPLE_PROPERTIES
+
 VM_PROPERTIES = VM_OBJECT_PROPERTIES + VM_SIMPLE_PROPERTIES
+
+ALL_PROPERTIES = VM_PROPERTIES + HOST_PROPERTIES + CLUSTER_PROPERTIES + DATASTORE_PROPERTIES
+
+SIMPLE_PROPERTIES_BY_RESOURCE_TYPE = {
+    'vm': VM_SIMPLE_PROPERTIES,
+    'host': HOST_SIMPLE_PROPERTIES,
+    'cluster': CLUSTER_SIMPLE_PROPERTIES,
+    'datastore': DATASTORE_SIMPLE_PROPERTIES,
+}
+
+PROPERTIES_BY_RESOURCE_TYPE = {
+    'vm': VM_PROPERTIES,
+    'host': HOST_PROPERTIES,
+    'cluster': CLUSTER_PROPERTIES,
+    'datastore': DATASTORE_PROPERTIES,
+}